Move events service to core/shared and use on home too
