+++ src/app/home/home/home.component.html
@@ -3,13 +3,8 @@
     <figure class="about-us__logo">
       <img src="assets/images/logo-gdg.svg" alt="logo" />
     </figure>
-    <div class="about-us__event-card">
-      <app-event-card
-        title="GDG DevFest Cochabamba 2018"
-        date="2018"
-        image="https://www.foot.com/wp-content/uploads/2017/03/placeholder.gif"
-        url="https://www.meetup.com/GDGCochabamba/events/253256943/">
-      </app-event-card>
+    <div class="about-us__event-card" *ngIf="mainEvent | async; let events">
+      <app-event-card [event]="events[0]"></app-event-card>
     </div>
   </div>
   <div class="hills city" >
@@ -53,13 +48,8 @@
         Codelabs, Meetups, Streaming, etc .... sobre diseño y desarrollo para web o móvil utilizando las herramientas y productos de Google.
       </p>
     </article>
-    <div class="about-do__event-card">
-      <app-event-card
-        title="GDG DevFest Cochabamba 2018"
-        date="2018"
-        image="https://lh3.googleusercontent.com/JKD1IzCYQWYRmuZUtpQtN63AL-AJ5x998xAzotX8-Tw6CLkxC8szUcN-MzHo8XqXvvtd=w1920-h1080-rw-no"
-        url="https://www.meetup.com/GDGCochabamba/events/253256943/">
-      </app-event-card>
+    <div class="about-do__event-card" *ngIf="randomEvent | async; let events">
+      <app-event-card [event]="events[0]"></app-event-card>
     </div>
     <div class="about-do__eventos">
       <a href="/eventos" class="section__link">
