+++ src/app/shared/event-card/event-card.component.html
@@ -2,7 +2,10 @@
   <figure class="event-card__image">
     <img [src]="event.image"/>
   </figure>
-  <div class="event-card__description" [ngClass]="[classDescription]" > 
+  <div
+    class="event-card__description"
+  >
+  <!-- [ngClass]="[classDescription]"  -->
     <div class="event-card__data">
       <h3 class="event-card__title">{{ event.title }}</h3>
       <h4 class="event-card__date">
